Deleted /src/test/java/org/opengis/cite/cdb10/cdbStructure/GSModel
