+++ src/main/java/org/opengis/cite/cdb10/GSModel/package-info.java
@@ -1,0 +1,1 @@
+package org.opengis.cite.cdb10.GSModel;
