--- src/main/java/org/opengis/cite/cdb10/cdbStructure/GSModel/package-info.java
@@ -1,1 +1,0 @@
-package org.opengis.cite.cdb10.cdbStructure.GSModel;
