+++ src/main/resources/org/opengis/cite/cdb10/testng.xml
@@ -13,9 +13,13 @@
   <test name="CDB Structure">
     <packages>
       <package name="org.opengis.cite.cdb10.cdbStructure" />
-      <package name="org.opengis.cite.cdb10.cdbStructure.GSModel" />
     </packages>
   </test>
+  <test name="CDB Structure - GSModel">
+    <packages>
+      <package name="org.opengis.cite.cdb10.GSModel" />
+    </packages>
+  </test>  
   <test name="Metadata And Versioning">
     <packages>
       <package name="org.opengis.cite.cdb10.metadataAndVersioning" />
